Add full defconfig to expose all options
As of the date of this commit, m8_defconfig & regen_m8_defconfig
will actually build with identical settings.  regen version is 4k
lines compared to less than 1k in the old.  The purpose of the regen
is to expose all of the options available to us. GCC fills in the
blanks for us with the m8_defconfig, and I'm not convinced that those
defaults it's picking for us are what we really want to use in all
cases.

m8_defconfig is definitely simpler though, so we'll keep both
for now. Only time will tell if the regen proves to be valuable.